% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemioCOVID.R
\name{alignMSA}
\alias{alignMSA}
\title{Aligning sample strends with reference covid strands with Muscle}
\usage{
alignMSA(seqset)
}
\arguments{
\item{seqset}{an alignment Set return by preAlign()}
}
\value{
Returns an alignment object
}
\description{
perform alignment between current COVID strands and one or more user-provided
genome sequence samples of COVID virus. Print the alignment.
}
\examples{
FA <- system.file("extdata", "samplefake.fasta", package = "epidemioCOVID")
seqs <- epidemioCOVID::readFASTA(FA)
msaSet <- preAlign(seqs, refseq = fakeref)
msa <- alignMSA(msaSet)

}
