% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BLOSUM62}
\alias{BLOSUM62}
\title{Attach the BLOSUM62 package from the Biostrings package}
\format{
data from BLOSUM62 for plotting
}
\usage{
BLOSUM62
}
\description{
A data for ploting
}
\examples{
\dontrun{
 BlOSUM62
}
}
\keyword{datasets}
