% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fakeref}
\alias{fakeref}
\title{A fake Gene of reference sequence of COVID19 virus from ncbi}
\format{
A characte string represents nucleotide chain:
}
\usage{
fakeref
}
\description{
A virus sequence generated from the FASTA format
}
\examples{
\dontrun{
 fakeref
}
}
\keyword{datasets}
