% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemioCOVID.R
\name{matchPer}
\alias{matchPer}
\title{Aligning sample strends with existing covid strands}
\usage{
matchPer(seqs)
}
\arguments{
\item{spl}{user-imported sample fasta data path name}

\item{ref}{user-imported sample fasta data path name}
}
\value{
Returns data.frame
$header char the FASTA header lines $seq double percentage of alignment
}
\description{
perform alignment between current COVID strands and one or more user-provided
genome sequence samples of COVID virus and identify the strand each sample
belongs to.
}
\examples{
FA <- system.file("extdata", "referseq.fasta", package = "epidemioCOVID")
seqs <- readFASTA(FA)
MP <- matchPer(seqs)


}
