% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemioCOVID.R
\name{msaPlot}
\alias{msaPlot}
\title{Ploting the alignment result to show conserved and diverged regions}
\usage{
msaPlot(msa)
}
\arguments{
\item{msa}{the alignment object generate by alignMSA()}
}
\value{
Returns plot showing the alinment regions
}
\description{
perform alignment between current COVID strands and one or more user-provided
genome sequence samples of COVID virus and identify the strand each sample
belongs to.
}
\examples{
FA <- system.file("extdata", "samplefake.fasta", package = "epidemioCOVID")
seqs <- epidemioCOVID::readFASTA(FA)
msaset <- preAlign(seqs, refseq = fakeref)
msa <- alignMSA(msaSet)
MP <- msaPlot(msa)

\dontrun{
FA <- system.file("extdata", "sample2.fasta", package = "epidemioCOVID")
seqs <- epidemioCOVID::readFASTA(FA)
MSA <- alignMSA(seqs)
MP <- msaPlot(MSA)
}

}
