% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemioCOVID.R
\name{mutsite.present}
\alias{mutsite.present}
\title{Representing Aligned sequence with muatation sites
perform alignment between current COVID strands and one or more user-provided
genome sequence samples of COVID virus and identify the strand each sample}
\usage{
mutsite.present(splfasta, start, end)
}
\arguments{
\item{spl}{of user-imported sample fasta data path name}
}
\value{
Returns the persentage of different nucleotides
}
\description{
Representing Aligned sequence with muatation sites
perform alignment between current COVID strands and one or more user-provided
genome sequence samples of COVID virus and identify the strand each sample
}
