% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemioCOVID.R
\name{preAlign}
\alias{preAlign}
\title{Prepare sample strends with reference covid strands for Alignment}
\usage{
preAlign(seqs, refseq = refseq)
}
\arguments{
\item{seqs}{user-imported sample multiple sequence fasta data path name}

\item{refseq}{refaseq we use, by defaul is the NC_045512.2}
}
\value{
Returns an alignment Set
}
\description{
Make a Mutiplealignment objects by Biostring package
}
\examples{
FA <- system.file("extdata", "samplefake.fasta", package = "epidemioCOVID")
seqs <- epidemioCOVID::readFASTA(FA)
msaSet <- preAlign(seqs, refseq = fakeref)

}
\references{
Boris Steipe. “Multiple Sequence Alignment.”
\href{http://steipe.biochemistry.utoronto.ca/bio/BIN-ALI-MSA.html}{Link}
}
