% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemioCOVID.R
\name{readFASTA}
\alias{readFASTA}
\title{read the FASTA format file}
\usage{
readFASTA(FA)
}
\arguments{
\item{FA}{Input file name (or text connection)}
}
\value{
Returns data.frame
$header char  the FASTA header lines $seq   char  the actual sequences
}
\description{
read the input in FASTA format
}
\examples{
FA <- system.file("extdata", "sample.fasta", package = "epidemioCOVID")
seqs <- readFASTA(FA)


}
\references{
Boris Steipe. “Introduction to R.”
\href{http://steipe.biochemistry.utoronto.ca/bio/RPR-Introduction.html}{Link}
}
