% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refseq}
\alias{refseq}
\title{Gene of reference sequence of COVID19 virus from ncbi}
\format{
A FASTA fomat sequence:
}
\source{
National Center for Biotechnology Information
}
\usage{
refseq
}
\description{
An FASTA format virus sequence
}
\examples{
\dontrun{
 refseq
}
}
\keyword{datasets}
