% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{splseq}
\alias{splseq}
\title{Gene of sample user-imported sequence of COVID19 virus from ncbi}
\format{
A FASTA fomat sequence:
}
\source{
National Center for Biotechnology Information
}
\usage{
splseq
}
\description{
An FASTA format virus sequence
}
\examples{
\dontrun{
 splseq
}
}
\keyword{datasets}
