% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemioCOVID.R
\name{matchPer}
\alias{matchPer}
\title{Aligning sample strends with existing covid strands}
\usage{
matchPer(spl, ref)
}
\arguments{
\item{spl}{user-imported sample fasta data path name}

\item{ref}{user-imported sample fasta data path name}
}
\value{
Returns the persentage of different nucleotides
}
\description{
perform alignment between current COVID strands and one or more user-provided
genome sequence samples of COVID virus and identify the strand each sample
belongs to with a graphical visualization of the mutation site in the sequence.
I would like to improve it so that this package can analyze multiple sequence
samples and propose the possible relation between samples to construct an
epidemiological transmission link if exits.
}
\examples{
load(file='data/refseq.rda')
load(file='data/splseq.rda')
matchPer(refseq, splseq)


}
