% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refseq}
\alias{refseq}
\title{Gene of reference sequence of COVID19 virus from ncbi}
\format{
A characte string represents nucleotide chain:
}
\source{
National Center for Biotechnology Information
}
\usage{
refseq
}
\description{
A virus sequence generated from the FASTA format, refID: NC_045512.2
}
\examples{
\dontrun{
 refseq
}
}
\keyword{datasets}
